Complete stake card reorganization
