+++ src/pages/Staking/StakeCard/StakeApplications/AuthorizeApplicationRow.tsx
